去除自定义HttpException1 HttpException类中新增构造 存入 code message 等信息
