+++ okgo/src/main/java/com/lzy/okgo/cache/policy/BaseCachePolicy.java
@@ -16,6 +16,7 @@
 package com.lzy.okgo.cache.policy;
 
 import android.graphics.Bitmap;
+import android.text.TextUtils;
 
 import com.lzy.okgo.OkGo;
 import com.lzy.okgo.cache.CacheEntity;
@@ -23,7 +24,6 @@ import com.lzy.okgo.cache.CacheMode;
 import com.lzy.okgo.callback.Callback;
 import com.lzy.okgo.db.CacheManager;
 import com.lzy.okgo.exception.HttpException;
-import com.lzy.okgo.exception.HttpException1;
 import com.lzy.okgo.model.Response;
 import com.lzy.okgo.request.base.Request;
 import com.lzy.okgo.utils.HeaderParser;
@@ -106,7 +106,7 @@ public abstract class BaseCachePolicy<T> implements CachePolicy<T> {
             if (responseCode == 404 || responseCode >= 500) {
                 String message = response.message();
                 String string = response.body().string();
-                return Response.error(false, rawCall, response, new HttpException1(responseCode, message, string));
+                return Response.error(false, rawCall, response, new HttpException(responseCode, message, string));
             }
 
             T body = request.getConverter().convertResponse(response);
@@ -156,7 +156,7 @@ public abstract class BaseCachePolicy<T> implements CachePolicy<T> {
                 if (responseCode == 404 || responseCode >= 500) {
                     String message = response.message();
                     String string = response.body().string();
-                    Response<T> error = Response.error(false, call, response, new HttpException1(responseCode,message,string));
+                    Response<T> error = Response.error(false, call, response, new HttpException(responseCode, message, string));
                     onError(error);
                     return;
                 }
